Include shader sizes in plugin size reporting. Add shader information to the cook metadata file. Update cook metadata dependency counts to uint32. Separate out plugin graph generation from usage. Add "other" size classification #rb arciel.rekman #rb fabian.giesen
[CL 27805498 by dan thompson in ue5-main branch]

+++ Engine/Plugins/Editor/AssetManagerEditor/Source/AssetManagerEditor/Private/TreeView/SAssetTableTreeView.cpp
@@ -1933,7 +1933,7 @@ void SAssetTableTreeView::RefreshAssets()
 			}
 
 			const int32 PluginIndex = AssetTable->GetIndexForPlugin(StoredPluginName);
-			for (uint16 DependencyIndexInMetadata = PluginEntry.DependencyIndexStart; DependencyIndexInMetadata < PluginEntry.DependencyIndexEnd; DependencyIndexInMetadata++)
+			for (uint32 DependencyIndexInMetadata = PluginEntry.DependencyIndexStart; DependencyIndexInMetadata < PluginEntry.DependencyIndexEnd; DependencyIndexInMetadata++)
 			{
 				const UE::Cook::FCookMetadataPluginEntry& DependentPluginEntry = PluginHierarchy.PluginsEnabledAtCook[PluginHierarchy.PluginDependencies[DependencyIndexInMetadata]];
 				const TCHAR* StoredReferencePluginName = AssetTable->StoreStr(DependentPluginEntry.Name);
